+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.8.3: October 20th, 2021
 * Update graph colors
 * New stat: total participants by status
 * Booking Form: new fields postalcode and city (need to add the fields to the GF forms also)